Add loading state for market
